correcting dtype issue in lstm 720_1 split test

+++ test_lstm_splts_720_1.py
@@ -46,7 +46,7 @@ for p_len in range(1, pred_len + 1):
     mae_by_pred_len[p_len] = 0.0
 
 for i, (x, y) in enumerate(single_loader(dataset)):
-    forecast = model(torch.tensor(np.array(x), device=device).unsqueeze(-1))[:, -pred_len:, :]
+    forecast = model(torch.tensor(np.array(x), device=device, dtype=torch.float32).unsqueeze(-1))[:, -pred_len:, :]
     
     y = torch.tensor(y).unsqueeze(-1)[:, -pred_len:, :].to(device=device)
